Initial Commit
+ collectionContainer.java
Stores the lists that contain the objects, and provides functions for sorting the lists, saving them to a file, and reading from a file.
+ Date.java
Class that stores information about a date in MM-DD-YYYY. Contains checks to ensure that the date stored is valid.
+ recordRelease.java
Class that stores information about an album.
+ Main.java
Currently blank, I've mainly been using this for testing. This file will likely be used later on for the GUI.

All code involving JSON uses the gson library provided by google